I added convertToString
So right now the two methods can convert from a string to a vector<char> and from a vector<char> to a string. For our use we don't need to cout the string/vector inside the conversion methods so if you need to use them just delete the for loop and the cout << s.