Operator.cpp
This is my attempt at the operations that Matt was working on earlier.
It includes an initialization constructor and a get Operand method.
However, when trying to perform the function, I first check to see if it
is an int in that spot or an e, pi, etc. It says that I can't add them
because the vector is of type input. How would i fix that?