Finds "rt:" in the string and grabs the n/x
This is not simplifying the root, just finding it in the string which will then be simplified (working on that now). I'm still going to add to this (convert from str1/str2 to int1/int2 - from there I will send int1/int2 to nthRootSolver), but this is most of it. This actual code will need to be in someone else's class